+++ modules/bridge/src/main/ts/ephox/bridge/components/dialog/Label.ts
@@ -1,22 +1,27 @@
-import { FieldProcessor } from '@ephox/boulder';
+import { FieldProcessor, FieldSchema } from '@ephox/boulder';
 
 import * as ComponentSchema from '../../core/ComponentSchema';
 import { BodyComponent, BodyComponentSpec } from './BodyComponent';
+
+type Alignment = 'start' | 'center' | 'end';
 
 export interface LabelSpec {
   type: 'label';
   label: string;
   items: BodyComponentSpec[];
+  align?: Alignment;
 }
 
 export interface Label {
   type: 'label';
   label: string;
   items: BodyComponent[];
+  align: Alignment;
 }
 
 export const createLabelFields = (itemsField: FieldProcessor): FieldProcessor[] => [
   ComponentSchema.type,
   ComponentSchema.label,
-  itemsField
+  itemsField,
+  FieldSchema.defaultedStringEnum('align', 'start', [ 'start', 'center', 'end' ])
 ];